requirements.txt
azure_storage==0.37.0
flet==0.7.4
numpy==1.23.5
Requests==2.30.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
TTS==0.14.0
twitchAPI==3.11.0
